+++ src/command/delete_database.rs
@@ -47,11 +47,7 @@ impl<'a> Command for DeleteDatabase<'a> {
     fn take_response(resp: hyper::client::Response, _state: Self::State) -> Result<Self::Output, Error> {
         match resp.status {
             hyper::status::StatusCode::Ok => transport::content_type_must_be_application_json(&resp.headers),
-            hyper::status::StatusCode::BadRequest => {
-                // The CouchDB spec says this status may also mean the document
-                // id has been "forgotten"--whatever that means!
-                Err(Error::InvalidDatabaseName(try!(ErrorResponse::from_reader(resp))))
-            }
+            hyper::status::StatusCode::BadRequest => Err(Error::BadRequest(try!(ErrorResponse::from_reader(resp)))),
             hyper::status::StatusCode::Unauthorized => Err(Error::Unauthorized(try!(ErrorResponse::from_reader(resp)))),
             hyper::status::StatusCode::NotFound => Err(Error::NotFound(Some(try!(ErrorResponse::from_reader(resp))))),
             _ => Err(Error::UnexpectedHttpStatus { got: resp.status }),