+++ src/command/post_to_database.rs
@@ -78,7 +78,7 @@ impl<'a, T> Command for PostToDatabase<'a, T> where T: 'a + serde::Serialize
                 let rev: Revision = content.rev.into();
                 Ok((rev, path))
             }
-            hyper::status::StatusCode::BadRequest => Err(Error::InvalidRequest(try!(ErrorResponse::from_reader(resp)))),
+            hyper::status::StatusCode::BadRequest => Err(Error::BadRequest(try!(ErrorResponse::from_reader(resp)))),
             hyper::status::StatusCode::Unauthorized => Err(Error::Unauthorized(try!(ErrorResponse::from_reader(resp)))),
             hyper::status::StatusCode::NotFound => Err(Error::NotFound(Some(try!(ErrorResponse::from_reader(resp))))),
             hyper::status::StatusCode::Conflict => {